+++ app/Http/Controllers/TransportesController.php
@@ -30,7 +30,10 @@ class TransportesController extends Controller
     public function show($id)
     {
         $transporte = Transporte::find($id);
-        return $transporte;
+        if($transporte!=NULL)
+            return $transporte;
+        else
+            return "El transporte con el id ingresado no existe o fue eliminado"; 
     }
 
     public function edit(Transporte $transporte)
@@ -49,7 +52,14 @@ class TransportesController extends Controller
     public function destroy($id)
     {
         $transporte = Transporte::find($id);
-        $transporte->delete();
-        return "Se ha eliminado el transporte de la DB";
+        if($transporte!=NULL)
+        {
+            $transporte->delete();
+            Transporte::destroy($id);
+            return "Se ha eliminado el transporte de la DB";
+        }
+        else
+            return "El transporte con el id ingresado no existe o fue eliminado"; 
+
     }
 }
